requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
docutils==0.18.1
easyocr==1.4.2
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
imageio==2.19.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
networkx==2.8.2
numpy==1.22.4
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.1.1
Pygments==2.12.0
pyparsing==3.0.9
pypiwin32==223
pytesseract==0.3.9
python-bidi==0.4.2
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
